Moved header and navbar to features folder and fixed imports to match
